pyproject.toml
[project]
name = "component-duckdb-transformation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "datadirtest",
    "duckdb>=1.3.1",
    "flake8>=7.3.0",
    "freezegun>=1.5.2",
    "keboola-component>=1.6.10",
    "keboola-http-client>=1.0.1",
    "keboola-utils>=1.1.0",
    "mock>=5.2.0",
    "pydantic>=2.11.7",
    "ruff>=0.12.7",
    "sqlglot>=25.0.0",
]

[tool.uv.sources]
datadirtest = { git = "https://github.com/keboola/datadirtest" }

[tool.ruff]
exclude = [
    ".git",
    "__pycache__",
    "tests",
    "example",
    "venv",
]
ignore = []
line-length = 120
select = [
    "E",
    "F",
    "W",
]
